Empirical computations

JenaAtomicCalculator.Empirical.AbstractEmpiricalApproximation — Type

abstract type Empirical.AbstractEmpiricalApproximation ... defines an abstract and a number of singleton types to deal with different line profiles, for instance, in a plasma.

+ GaussianProfile       ... assumes a Gaussian line profile L^Gaussian (omega)
+ LorentzianProfile     ... assumes a Lorentzian line profile L^Lorentzian (omega)

JenaAtomicCalculator.Empirical.Computation — Type

struct Empirical.Computation ... defines a type for an empirical computation of various ionization and charge exchange processes.

+ name                   ::String                           ... A name associated to the computation.
+ nuclearModel           ::Nuclear.Model                    ... Model, charge and parameters of the nucleus.
+ grid                   ::Radial.Grid                      ... The radial grid to be used for the computation.
+ configs                ::Array{Configuration,1}           ... A list of non-relativistic configurations.
+ settings               ::Basics.AbstractEmpiricalSettings ... Provides the settings for the selected computations.

JenaAtomicCalculator.Empirical.Computation — Method

Empirical.Computation(comp::Empirical.Computation;

name=..,                nuclearModel=..,            grid=..,                    configs=..,                   settings=..,  
printout::Bool=false)
                
... constructor for modifying the given Empirical.Computation by 'overwriting' the previously selected parameters.

JenaAtomicCalculator.Empirical.Computation — Method

Empirical.Computation() ... constructor for an 'empty' instance::Empirical.Computation.

JenaAtomicCalculator.Empirical.GivenEinsteinA — Type

struct GivenEinsteinA <: Empirical.AbstractEmpiricalApproximation ... to communicate an externally given Einstein-A value, together with the multipole and energy of the transition i –> f.

+ multipole   ::EmMultipole     ... Multipole of the transition.
+ energy      ::Float64         ... Energy of transition.
+ rate        ::Float64         ... Einstein-A value.

JenaAtomicCalculator.Basics.perform — Method

Basics.perform(computation::Empirical.Computation) ... to set-up and perform an empirical computation that starts from a given nuclear model and set of configurations, and which is mainly controlled by its settings. The results are printed to screen but nothing is returned otherwise.

Basics.perform(computation::Empirical.Computation; output=true) ... to perform the same but to return the complete output in a dictionary; the particular output depends on the kind and specifications of the empirical computation but can easily accessed by the keys of this dictionary.

JenaAtomicCalculator.Empirical.bindingEnergy — Method

Empirical.bindingEnergy(Z::Int64, sh::Shell; data::PeriodicTable.AbstractEnergyData=Williams2000() ) ... to estimate the binding) energy of the given shell for a neutral atom; these binding energies are taken from semi-empirical tabulations due to data by different data sets. An energy::Float64 > 0. is returned. ...... Perhaps, introduce an optional argument scaleLast::Bool to scale values, if no explicit binding energies are available.

JenaAtomicCalculator.Empirical.bindingEnergy — Method

Empirical.bindingEnergy(Z::Int64, subsh::Subshell, conf::Configuration; data::PeriodicTable.AbstractEnergyData=Williams2000()) ... to evaluate the binding energy of a subshell electron if part of the given configuration. The tabulated binding energies refer to the neutral atom; this procedure assumes that each missing electron (compared with the neutral system) adds 0.3 / n Hartree to the binding energy, where n is the principal quantum number of the subshell. It also requires that the shell associated with subsh is occupied in the given configuration. An energy::Float64 is returned.

JenaAtomicCalculator.Empirical.bindingEnergy — Method

Empirical.bindingEnergy(Z::Int64, subsh::Subshell; data::PeriodicTable.AbstractEnergyData=Williams2000() ) ... to estimate the binding) energy of the given subshell for a neutral atom; these binding energies are taken from semi-empirical tabulations due to data by different data sets. An energy::Float64 > 0. is returned.

JenaAtomicCalculator.Empirical.ionizationPotential — Method

Empirical.ionizationPotential(Z::Int64, conf::Configuration) ... to evaluate the ionization potential I_p that is needed to ionize a valence electron from the given configuration. An energy::Float64 is returned.

JenaAtomicCalculator.Empirical.meanCharge — Method

Empirical.meanCharge(Z::Int64, subshell::Subshell, conf::Configuration) ... to estimate the mean charge Z as seen by an electron in subshell for an ion with the given configuation. This charge is obtained by inverting the hydrogenic relation bE = Z^(eff)^2 / (2 n^2) for the binding energy bE > 0. of the given subshell, i.e. Z^(eff) = n sqrt(2 bE). A charge::Float64 is returned.

JenaAtomicCalculator.Empirical.photodeexcitationPlasmaRatePerIon — Method

Empirical.photodeexcitationPlasmaRatePerIon(dist::Distribution.AbstractPhotonDistribution, iConf::Configuration, fConf::Configuration; approx::Empirical.AbstractEmpiricalApproximation=UsingJAC(), printout::Bool=false)

... to estimate the total photodeexcitation plasma rate R^(PD: total, per ion) (T; i -> f) for a transition from 
    iConf -> fConf by applying a given photon distribution dist. A rate::Float64 is returned.

JenaAtomicCalculator.Empirical.photoemissionEinsteinA — Method

Empirical.photoemissionEinsteinA(iConf::Configuration, fConf::Configuration, approx::Empirical.ScaledHydrogenic; printout::Bool=false, data::PeriodicTable.AbstractEnergyData=PeriodicTable.Williams2000()) ... to estimate empirically the Einstein-A value for a transition from iConf -> fConf. The transition energy is obtained as the difference of the binding energies of the two shells involved, cf. Empirical.scaledBindingEnergy(), while the A-value follows from a hydrogenic scaling with an effective nuclear charge. A named triple (multipole::EmMultipole=, energy::Float64=, rate::Float64=) is returned.

JenaAtomicCalculator.Empirical.photoemissionEinsteinA — Method

Empirical.photoemissionEinsteinA(iConf::Configuration, fConf::Configuration, approx::Empirical.UsingJAC; printout::Bool=false) ... to estimate empirically the Einstein-A value for a transition from iConf -> fConf by using simple JAC computations of transition and Einstein rates. A named triple (multipole::EmMultipole=, energy::Float64=, rate::Float64=) is returned.

JenaAtomicCalculator.Empirical.photoemissionEinsteinA — Method

Empirical.photoemissionEinsteinA(approx::Empirical.GivenEinsteinA; printout::Bool=false) ... to simply return the Einstein-A value for the given transition. A named (multipole::EmMultipole=, energy::Float64=, rate::Float64=) is returned.

JenaAtomicCalculator.Empirical.photoexcitationPlasmaRatePerIon — Method

Empirical.photoexcitationPlasmaRatePerIon(dist::Distribution.AbstractPhotonDistribution, iConf::Configuration, fConf::Configuration; approx::Empirical.AbstractEmpiricalApproximation=UsingJAC(), printout::Bool=false)

... to estimate the photoexcitation plasma rate per ion R^(PX: per ion) (T; i -> f) for a transition from 
    iConf -> fConf by applying a given photon distribution dist. A rate::Float64 is returned.

JenaAtomicCalculator.Empirical.photoionizationCrossSection — Method

Empirical.photoionizationCrossSection(omega::Float64, iConf::Configuration, fConf::Configuration; approx::Empirical.AbstractEmpiricalApproximation=ScaledHydrogenic(), printout::Bool=false)

... to estimate empirically the PI cross section for a transition from iConf -> fConf by applying some simple approximation 
    as determined by approx. For printout=true, basic information are printed about the input parameters, approximation 
    as well as the results in user-defined units. A cs::Float64 [in a.u.] is returned.

JenaAtomicCalculator.Empirical.photoionizationCrossSection — Method

Empirical.photoionizationCrossSection(omegas::Array{Float64,1}, iConf::Configuration, fConf::Configuration, approx::Empirical.ScaledHydrogenic; printout::Bool=false, data::PeriodicTable.AbstractEnergyData=PeriodicTable.Williams2000()) ... to estimate empirically the photoionization cross sections for a transition from iConf -> fConf by using the binding energy (ionization potential) of the ionized shell, cf. Empirical.scaledBindingEnergy(), and Kramer's (1923) empirical formula. A css::Array{Float64,1} [a.u.] is returned.

JenaAtomicCalculator.Empirical.photoionizationCrossSection — Method

Empirical.photoionizationCrossSection(omegas::Array{Float64,1}, iConf::Configuration, fConf::Configuration, approx::Empirical.UsingJAC; printout::Bool=false) ... to estimate the photoionization cross section for a transition from iConf -> fConf from mean-field orbitals and single-electron E1 amplitudes. A css::Array{Float64,1} [a.u.] is returned.

    !! NOT VALIDATED; this method presently raises an error. !!
    Its cross sections *grow* with the photon energy above the threshold, whereas any PI cross section has to fall:
    for Ne 2p it yields 5.7 Mb at 25 eV but 92 Mb at 100 eV, where the experimental value has dropped to ~1 Mb.
    The prefactor 8 pi^3 / (alpha omega) contradicts the defining relation
    sigma^(PI)(omega; i -> f) = 4 pi^2 alpha omega sum_lambda |<f| D_lambda |i>|^2,
    i.e. both the alpha- and the omega-dependence appear inverted. Since MabEmissionJohnsony provides an *emission*
    amplitude, its normalization has to be settled before this prefactor can be re-derived; cf. the (correct)
    treatment in module-PhotoIonization.jl. The body below is retained unchanged for that repair.

JenaAtomicCalculator.Empirical.photoionizationPlasmaRatePerIon — Method

Empirical.photoionizationPlasmaRatePerIon(dist::Distribution.AbstractPhotonDistribution, iConf::Configuration, fConf::Configuration; approx::Empirical.AbstractEmpiricalApproximation=ScaledHydrogenic(), printout::Bool=false, data::PeriodicTable.AbstractEnergyData=PeriodicTable.Williams2000()) ... to estimate empirically the photoionization plasma rate per ion R^(PI: per ion) (T; i -> f) for a transition from iConf -> fConf and a photon field dist at the radiation temperature T. A rate::Float64 [a.u.] is returned.

    Note: In contrast to the photorecombination coefficient alpha^(PR) [cm^3/s], this quantity is a *rate* [1/s] and
          not a rate coefficient. The convolution R^(PI: per ion) = int d(omega) n(omega; T) c sigma^(PI)(omega)
          already contains the photon number density of the radiation field and, hence, needs no further
          multiplication by a density; it is the number of photoionization events per ion and per unit time.
          The free-electron density, in contrast, does not enter the PI process at all.

JenaAtomicCalculator.Empirical.photorecombinationCrossSection — Method

Empirical.photorecombinationCrossSection(energy::Float64, iConf::Configuration, fConf::Configuration; approx::Empirical.AbstractEmpiricalApproximation=ScaledHydrogenic(), printout::Bool=false)

... to estimate empirically the PR cross section for a transition from iConf -> fConf by applying some simple approximation 
    as determined by approx. For printout=true, basic information are printed about the input parameters, approximation 
    as well as the results in user-defined units. A cs::Float64 [in a.u.] is returned.

JenaAtomicCalculator.Empirical.photorecombinationCrossSection — Method

Empirical.photorecombinationCrossSection(energies::Array{Float64,1}, iConf::Configuration, fConf::Configuration, approx::Empirical.ScaledHydrogenic; printout::Bool=false, data::PeriodicTable.AbstractEnergyData=PeriodicTable.Williams2000()) ... to estimate empirically the (spontaneous) photorecombination cross section for a transition from iConf -> fConf by using the Einstein-Milne relation and the binding energy (ionization potential) of the captured electron, cf. Empirical.scaledBindingEnergy(), together with Kramer's (1923) empirical formula. A css::Array{Float64,1} [a.u.] is returned.

JenaAtomicCalculator.Empirical.photorecombinationCrossSection — Method

Empirical.photorecombinationCrossSection(energies::Array{Float64,1}, iConf::Configuration, fConf::Configuration, approx::Empirical.UsingJAC; printout::Bool=false) ... to estimate the (spontaneous) photorecombination cross section for a transition from iConf -> fConf by applying the Einstein-Milne relation to mean-field PI cross sections; the binding energy of the captured electron is taken from the mean-field orbital of fConf. A css::Array{Float64,1} [a.u.] is returned.

    !! NOT VALIDATED; this method presently raises an error. !!
    Its Einstein-Milne relation and an undefined variable have been corrected, but the method rests on
    Empirical.photoionizationCrossSection(..., UsingJAC), whose cross sections are themselves unphysical. As a
    consequence, these PR cross sections *grow* with the electron energy instead of exhibiting the characteristic
    divergence sigma^(PR) ~ 1/eps as eps -> 0. Fixing the PI normalization above will repair this method as well.

JenaAtomicCalculator.Empirical.photorecombinationPlasmaAlpha — Method

Empirical.photorecombinationPlasmaAlpha(dist::Distribution.AbstractElectronDistribution, iConf::Configuration, fConf::Configuration; approx::Empirical.AbstractEmpiricalApproximation=ScaledHydrogenic(), printout::Bool=false) ... to estimate empirically the spontaneous PR plasma rate coefficient alpha^(PR: spontaneous) for a transition from iConf -> fConf by applying some simple approximation as determined by approx. For printout=true, basic information are printed about the approximation as well as the results. An alpha::Float64 is returned. This is a short-cut to the method above for a vanishing (vacuum) photon field.

JenaAtomicCalculator.Empirical.photorecombinationPlasmaAlpha — Method

Empirical.photorecombinationPlasmaAlpha(eDist::Distribution.AbstractElectronDistribution, pDist::Distribution.AbstractPhotonDistribution, iConf::Configuration, fConf::Configuration; approx::Empirical.AbstractEmpiricalApproximation=ScaledHydrogenic(), printout::Bool=false, data::PeriodicTable.AbstractEnergyData=PeriodicTable.Williams2000()) ... to estimate empirically the total PR plasma rate coefficient alpha^(PR: total), i.e. the sum of the spontaneous and the stimulated contribution, for a transition from iConf -> fConf. The free electrons follow the distribution eDist at the electron temperature T_e, while the ambient radiation field is described by pDist at the (independent) radiation temperature T. An alpha::Float64 [a.u.] is returned.

    Note: The stimulated enhancement [1 + nbar(omega)] is evaluated *inside* the integral over the electron energies.
          Photorecombination is a bound-free process, for which the emitted photon energy omega = eps + bEnergy is not
          a single number but varies across the whole electron distribution; there is, therefore, no unique omega at
          which a common factor [1 + nbar] could be taken out of the integral. Such a factorization is exact only for
          bound-bound processes (PD, PX), where omega = E_f - E_i is fixed. It would introduce an error that grows with
          the departure of the radiation field from the electron temperature, i.e. just in the non-LTE regime.
          For pDist = PhotonVacuumField, nbar = 0 and the spontaneous coefficient is recovered.

JenaAtomicCalculator.Empirical.scaledBindingEnergy — Method

Empirical.scaledBindingEnergy(Z::Float64, sh::Shell, conf::Configuration, data::PeriodicTable.AbstractEnergyData) ... to provide the binding energy of an electron in shell sh of the given configuration conf; this energy scales all hydrogenic (Kramers) estimates below and, hence, fixes the photoionization threshold as well as the transition energies. The semi-empirical binding energies of the neutral atom are taken from the tabulation data whenever this data set covers Z and sh. If no tabulated value is available, a Slater-screened hydrogenic estimate bE = Z^(eff)^2 / (2 n^2) is applied instead and a warning is issued; Z^(eff) then follows from Slater's (1930) rules for the remaining electrons of conf. An energy::Float64 > 0. [a.u.] is returned.

    Note: A pure hydrogenic estimate is quite unreliable for valence shells; for neutral Ne, for instance, it places
          the 2p threshold near 85 eV, while the tabulated (true) binding energy is 21.6 eV. The tabulated values
          should therefore be preferred whenever they are available.

JenaAtomicCalculator.Empirical.stoppingPowerElectronBethe1931 — Method

Empirical.stoppingPowerElectronBethe1931(energy::Float64, nRatio::Float64, omegaPlasma::Float64; printout::Bool=false) ... to evaluate the stopping power - d E / dx of an electron with kinetic energy in a plasma with nRatio = nelectrons / nions and for the plasma frequency omegaPlasma. A sPower::Float64 is returned. If printout = true, the value is printed in user-defined values along with short other explanations.

JenaAtomicCalculator.Empirical.stoppingPowerElectronBohr1913 — Method

Empirical.stoppingPowerElectronBohr1913(energy::Float64, nRatio::Float64, omegaPlasma::Float64; printout::Bool=false) ... to evaluate the stopping power - d E / dx of an electron with kinetic energy in a plasma with nRatio = nelectrons / nions and for the plasma frequency omegaPlasma. A sPower::Float64 is returned. If printout = true, the value is printed in user-defined values along with short other explanations.

JenaAtomicCalculator.Empirical.totalEnergy — Method

Empirical.totalEnergy(Z::Int64, conf::Configuration; data::PeriodicTable.AbstractEnergyData=Nist2025() ) ... to estimate the total energy of the given configuration for an atom or ions with nuclear charge Z; this total energy is determined either from the (negative) sum of all binding energies or the (negative) sum of ionization potentials that are needed to remove the remaining electrons from the configuration. A warning is issued if electrons occur in excited shells with regard to the ground configuration of the neutral atom. An energy::Float64 is returned.

    Note: This is an estimate of the total energy which need not to be so accurate.

Semi-empirical estimates

JenaAtomicCalculator.Semiempirical.estimate — Method

Semiempirical.estimate(::EstimateBindingEnergyLarkins1977, Z::Int64, conf::Configuration) ... provides an approximate total binding energy of a configuration from Larkins (1977); no relaxation effects between hole states are included; an energy::Float64 in Hartree is returned.

JenaAtomicCalculator.Semiempirical.estimate — Method

Semiempirical.estimate(::EstimateBindingEnergyLarkins1977, Z::Int64, sh::Subshell) ... provides the binding energy of a subshell electron from the Larkins (1977) tabulation; an energy::Float64 in Hartree is returned.

JenaAtomicCalculator.Semiempirical.estimate — Method

Semiempirical.estimate(::EstimateBindingEnergyNist2025, Z::Int64, q::Int64) ... returns the q-th successive ionization potential for element Z from the NIST (2025) database, https://physics.nist.gov/PhysRefData/ASD/ionEnergy.html; q = 1 gives the first ionization potential (outermost/valence electron of the neutral atom), q = 2 the second, and so on. Covers Z = 1..90; an energy::Float64 in Hartree is returned.

JenaAtomicCalculator.Semiempirical.estimate — Method

Semiempirical.estimate(::EstimateBindingEnergyWilliams2000, Z::Int64, conf::Configuration) ... provides an approximate total binding energy of a configuration from Williams et al. (2000); no relaxation effects between hole states are included; an energy::Float64 in Hartree is returned.

JenaAtomicCalculator.Semiempirical.estimate — Method

Semiempirical.estimate(::EstimateBindingEnergyWilliams2000, Z::Int64, sh::Subshell) ... provides the binding energy of a subshell electron from the Williams et al. (2000) tabulation, https://userweb.jlab.org/~gwyn/ebindene.html; an energy::Float64 in Hartree is returned.

JenaAtomicCalculator.Semiempirical.estimate — Method

Semiempirical.estimate(::EstimateBindingEnergyXrayDataBooklet, Z::Int64, conf::Configuration) ... provides an approximate total binding energy of a configuration from the X-ray Data Booklet tabulation; no relaxation effects between hole states are included; an energy::Float64 in Hartree is returned.

JenaAtomicCalculator.Semiempirical.estimate — Method

Semiempirical.estimate(::EstimateBindingEnergyXrayDataBooklet, Z::Int64, sh::Subshell) ... provides the binding energy of a subshell electron from the X-ray Data Booklet tabulation; covers 1s1/2 through 6p3/2 (24 subshells); an energy::Float64 in Hartree is returned.

JenaAtomicCalculator.Semiempirical.estimate — Method

Semiempirical.estimate(::EstimateIonizationPotentialInnerShell, sh::Subshell, Z::Int64) ... estimates the ionization potential = binding energy of an electron in the given subshell; values are taken from Williams et al. (2000); covers 1s1/2 through 4p3/2 (12 subshells); an energy value::Float64 in Hartree is returned.

JenaAtomicCalculator.Semiempirical.estimateBindingEnergies — Method

Semiempirical.estimateBindingEnergies(Z::Float64, coreConf::Configuration, nRange::UnitRange{Int64}, l::Int64) ... to estimate the binding energies of the high-n shells with n = nRange and orbital angular momentum l for a (multiply-charged) ion with nuclear charge Z and core configuration coreConfiguration; Slater's (1930) screening rules are used to compute Z_eff for each (n, l) shell. A list of energies::Array{Float64} [in Hartree] is returned.

JenaAtomicCalculator.Semiempirical.estimateBindingEnergies — Method

Semiempirical.estimateBindingEnergies(Z::Float64, coreConf::Configuration, nRange::UnitRange{Int64}) ... to estimate the binding energies of the high-n shells with n = nRange for a (multiply-charged) ion with nuclear charge Z and core configuration coreConfiguration. A list of energies::Array{Float64} [in Hartree] is returned.

JenaAtomicCalculator.Semiempirical.estimateRadialExpectation — Method

Semiempirical.estimateRadialExpectation(Z::Float64, coreConf::Configuration, nRange::UnitRange{Int64}, l::Int64) ... to estimate the radial extent (r-expectation value) of the high-n shells with n = nRange and orbital angular momentum l for a (multiply-charged) ion with nuclear charge Z and core configuration coreConfiguration. A list of radii::Array{Float64} [in a_o] is returned.

JenaAtomicCalculator.Semiempirical.estimateSlaterZeff — Method

Semiempirical.estimateSlaterZeff(Z::Float64, conf::Configuration, sh::Subshell) ... estimates the effective nuclear charge Z_eff for a single electron in subshell sh, given the remaining electrons in conf and the nuclear charge Z. Slater's (1930) empirical screening rules are applied: same-group electrons contribute σ = 0.35 (0.30 for [1s]), electrons with n' = n−1 contribute σ = 0.85 (sp targets only), and all inner electrons contribute σ = 1.00. For d and f electrons all inner electrons contribute σ = 1.00. The configuration conf must represent the other N−1 electrons only (not include the target electron itself); a Zeff::Float64 is returned.