Atomic form factors

JenaAtomicCalculator.FormFactor.Outcome — Type

struct FormFactor.Outcome ... defines a type to keep the outcome of a form-factor computation, such as the standard form factor as well other results.

+ level                     ::Level              ... Atomic level to which the outcome refers to.
+ qValues                   ::Array{Float64,1}   ... momentum transfer |q|.
+ standardFs                ::Array{Float64,1}   ... standard atomic form factor F(q) for an (assumed) spherical charge distr.
+ modifiedFs                ::Array{Float64,1}   ... modified atomic form factor F(q) for such a charge distribution.

JenaAtomicCalculator.FormFactor.Outcome — Method

FormFactor.Outcome() ... constructor for an empty instance of FormFactor.Outcome for the computation of atomic form factors.

JenaAtomicCalculator.FormFactor.Settings — Type

struct FormFactor.Settings <: AbstractPropertySettings ... defines a type for the details and parameters of computing alpha-variation parameters.

+ qList                    ::Array{Float64,1} ... List of q-values in [a.u.]
+ printBefore              ::Bool             ... True if a list of selected levels is printed before the  
                                                    actual computations start. 
+ levelSelection           ::LevelSelection   ... Specifies the selected levels, if any.

JenaAtomicCalculator.FormFactor.Settings — Method

FormFactor.Settings(set::FormFactor.Settings;

    qList=.., printBefore=.., levelSelection=..)

... keyword copy-constructor for re-defining selected values of a settings::FormFactor.Settings.

JenaAtomicCalculator.FormFactor.Settings — Method

FormFactor.Settings() ... constructor for an empty instance of FormFactor.Settings for the computation of atomic form factors.

JenaAtomicCalculator.FormFactor.amplitude — Method

FormFactor.amplitude(q::Float64, finalLevel::Level, initialLevel::Level, grid::Radial.Grid; display::Bool=false) ... to compute the momentum transfer amplitude <(alphaf Jf, kappa) Ji || T^(1) || alphai J_i> for the given final and initial level. A value::ComplexF64 is returned.

JenaAtomicCalculator.FormFactor.computeAmplitudesProperties — Method

FormFactor.computeAmplitudesProperties(outcome::FormFactor.Outcome, nm::Nuclear.Model, grid::Radial.Grid, settings::FormFactor.Settings) ... to compute all form factors, etc. for the given outcome; a newOutcome::FormFactor.Outcome is returned.

JenaAtomicCalculator.FormFactor.computeOutcomes — Method

FormFactor.computeOutcomes(multiplet::Multiplet, nm::Nuclear.Model, grid::Radial.Grid, settings::FormFactor.Settings; output=true) ... to compute (as selected) the alpha-variation parameters for the levels of the given multiplet and as specified by the given settings. The results are printed in neat tables to screen but nothing is returned otherwise.

JenaAtomicCalculator.FormFactor.determineOutcomes — Method

FormFactor.determineOutcomes(multiplet::Multiplet, settings::FormFactor.Settings) ... to determine a list of Outcomes's for the computation of the alpha-variation parameters for the given multiplet. It takes into account the particular selections and settings. An Array{FormFactor.Outcome,1} is returned. Apart from the level specification, all physical properties are set to zero during the initialization process.

JenaAtomicCalculator.FormFactor.displayOutcomes — Method

FormFactor.displayOutcomes(stream::IO, outcomes::Array{FormFactor.Outcome,1}) ... to display a list of levels that have been selected for the computations. A small neat table of all selected levels and their energies is printed but nothing is returned otherwise.

JenaAtomicCalculator.FormFactor.displayResults — Method

FormFactor.displayResults(stream::IO, outcomes::Array{FormFactor.Outcome,1}) ... to display the energies, M_ms and F-parameters, etc. for the selected levels. A neat table is printed but nothing is returned otherwise.

JenaAtomicCalculator.FormFactor.modifiedF — Method

FormFactor.modifiedF(q::Float64, level::Level, grid::Radial.Grid) ... to compute the (real) Fourier transform of the (spherically-symmetric) charge distribution but including a correction due to the local (DFS) potential, relative to the rest mass of the electron. A value::Float64 is returned.

JenaAtomicCalculator.FormFactor.standardF — Method

FormFactor.standardF(q::Float64, level::Level, grid::Radial.Grid) ... to compute the (real) Fourier transform of the (spherically-symmetric) charge distribution. A value::Float64 is returned.

Hyperfine coefficients and energy splittings

JenaAtomicCalculator.Hfs.HfBasisVector — Type

struct Hfs.HfBasisVector ... defines a type for a hyperfine basis vector that enables one to think about and deal with individual hyperfine levels. These hyperfine levels have representations with regard to (tensor) product states, which are formed from a set of isomeric states as well as a set of electronic ASF level J. Hyperfine basis vectors are obtained from diagonalizing the hyperfine Hamiltonian in a tensor basis of isomeric + ASF states.

+ F         ::AngularJ64        ... Total angular momentum F
+ parity    ::Parity            ... Total parity of the basis vector = nuclear x electronic parity.
+ isomer    ::Nuclear.Isomer    ... Isomeric state of the nucleus.
+ LevelJ    ::Basis             ... Electronic level that is part of the electronic basis.

There is no need to introduce a type HfBasis since such a hfBasis = Hfs.HfBasisVector[...] can be readily formed at all 
occurrences. The IJF_Vector's and     HfBasisVector's have different advantages.

JenaAtomicCalculator.Hfs.HfBasisVector — Method

Hfs.HfBasisVector() ... constructor for an empty instance of HfBasisVector`.

JenaAtomicCalculator.Hfs.HfLevel — Type

struct Hfs.HfLevel ... defines a type for HfLevel with a representation that refers to a product basis of isomeric and ASF states; the HfLevel is the pendant to a (electronic) Level/state, if hyperfine-resolved transitions are considered. Each hyperfine level has a representation mc that refers to the hfBasis, and which contains all information about the representation of the underlying nuclear and electronic basis states. The electronic basis is formed by a selected set of ASF, typically taken from some (electronic) multiplet. In contrast to a pure (electronic) IJF_Basis, the use of ASF simplifies the interpretation of physical findings but cannot reduce the computational effort (perhaps, even results in slightly increase the computational effort). A HfLevel is defined by:

+ F              ::AngularJ64               ... Total angular momentum F.
+ M              ::AngularM64               ... Total projection M, only defined if a particular magnetic sublevel is referred to.
+ parity         ::Parity                   ... Parity of the level which corresponds to the electronic system.
+ energy         ::Float64                  ... energy
+ hfBbasis       ::Array{HfBasisVector,1}   ... the product basis nuclear (isomeric) state x selected ASF.
+ mc             ::Vector{Float64}          ... Vector of mixing coefficients w.r.t hfBasis.

JenaAtomicCalculator.Hfs.HfLevel — Method

Hfs.HfLevel() ... constructor for an empty instance of HfLevel.

JenaAtomicCalculator.Hfs.HfMultiplet — Type

struct Hfs.HfMultiplet ... defines a type for a multiplet of hyperfine levels (HfLevel's) which are based on a (tensor) product basis of nuclear (isomeric) x ASF states.

+ name     ::String                ... A name associated to the multiplet.
+ hfLevels ::Array{HfLevel,1}      ... List of hyperfine levels (HfLevel's)

JenaAtomicCalculator.Hfs.HfMultiplet — Method

Hfs.HfMultiplet() ... constructor for an empty instance of Hfs.HfMultiplet.

JenaAtomicCalculator.Hfs.IJF_Vector — Type

struct Hfs.IJF_Vector ... defines a type for a IJF-coupled basis vector, here based on an ASF. Following IJF-coupling, an IJF basis vector is always the (tensor) product state of an isomeric state x electronic state. IJF-coupled basis vector are needed and obtained from the diagonalization of the (electronic) HFS Hamiltonian within a given CSF basis.

+ F         ::AngularJ64        ... Total angular momentum F
+ parity    ::Parity            ... Total parity of the basis vector = nuclear x electronic parity.
+ isomer    ::Nuclear.Isomer    ... Isomeric state of the nucleus.
+ csf       ::ManyElectron.CsfR ... Electronic CSF state.
+ basisJ    ::Basis             ... Electronic CSF basis to which the csf refer to.

There is no need to introduce a type IJF_Basis since such a ijfBasis = Hfs.IJF_Vector[...] can be readily formed at all 
occurrences.

JenaAtomicCalculator.Hfs.IJF_Vector — Method

Hfs.IJF_Vector() ... constructor for an empty instance of IJF_Vector.

JenaAtomicCalculator.Hfs.InteractionMatrix — Type

struct Hfs.InteractionMatrix ... defines a type for storing the T^1 and T^2 interaction matrices for a given basis.

+ calcM1   ::Bool               ... true, if the matrixM1 has been calculated and false otherwise.
+ calcE2   ::Bool               ... true, if the matrixE2 has been calculated and false otherwise.
+ calcM3   ::Bool               ... true, if the matrixM3 has been calculated and false otherwise.
+ matrixM1 ::Array{Float64,2}   ... T^M1 interaction matrix
+ matrixE2 ::Array{Float64,2}   ... T^E2 interaction matrix
+ matrixM3 ::Array{Float64,2}   ... T^M3 interaction matrix

JenaAtomicCalculator.Hfs.InteractionMatrix — Method

Hfs.InteractionMatrix() ... constructor for an empty instance of InteractionMatrix.

JenaAtomicCalculator.Hfs.Outcome — Type

struct Hfs.Outcome ... defines a type to keep the outcome of a HFS computation, such as the HFS A and B coefficients as well other results.

+ Jlevel                    ::Level            ... Atomic level to which the outcome refers to.
+ gJ                        ::Float64          ... Lande's g_J factor of the level.
+ AIoverMu                  ::Float64          ... HFS A * I / mu value.
+ BoverQ                    ::Float64          ... HFS B / Q value
+ CoverOmega                ::Float64          ... HFS C / Omega value
+ amplitudeM1               ::Complex{Float64} ... M1 amplitude
+ amplitudeE2               ::Complex{Float64} ... E2 amplitude
+ amplitudeM3               ::Complex{Float64} ... M3 amplitude
+ nuclearI                  ::AngularJ64       ... nuclear spin
+ hfsMultiplet              ::HfMultiplet      ... Multiplet of HfLevel's as associated with the JLevel.

JenaAtomicCalculator.Hfs.Outcome — Method

Hfs.Outcome() ... constructor for an empty instance of Hfs.Outcome for the computation of HFS properties.

JenaAtomicCalculator.Hfs.Settings — Type

struct Settings <: AbstractPropertySettings ... defines a type for the details and parameters of computing HFS A and B coefficients.

+ calcM1                    ::Bool             ... True if T^M1-amplitudes (HFS A values) need to be calculated, and false otherwise.
+ calcE2                    ::Bool             ... True if T^E2-amplitudes (HFS B values) need to be calculated, and false otherwise.
+ calcM3                    ::Bool             ... True if T^M3-amplitudes (HFS C values) need to be calculated, and false otherwise.
+ calcNondiagonal           ::Bool             
    ... True if also (non-)diagonal hyperfine amplitudes are to be calculated and printed, and false otherwise.
+ calcHfMultiplet           ::Bool             
    ... True if representation of all hyperfine levels need to be generated in the IJF-coupled ASF basis;
        only the nuclear spin and the selected atomic levels are considered in the basis. false otherwise.
+ printBefore               ::Bool             ... True if a list of selected levels is printed before the actual computations start. 
+ levelSelection            ::LevelSelection   ... Specifies the selected levels, if any.

JenaAtomicCalculator.Hfs.Settings — Method

Hfs.Settings(set::Hfs.Settings;

    calcM1=.., calcE2=.., calcM3=.., calcNondiagonal=.., calcHfMultiplet=.., printBefore=.., levelSelection=..)

... keyword copy-constructor for re-defining selected values of a settings::Hfs.Settings.

JenaAtomicCalculator.Hfs.Settings — Method

Hfs.Settings(; calcM1::Bool=true, calcE2::Bool=false, calcM3::Bool=false, calcNondiagonal::Bool=false, calcHfMultiplet::Bool=false, printBefore::Bool=false, levelSelection::LevelSelection=LevelSelection()) ... keyword constructor to create a Hfs.Settings with selected non-default values.

JenaAtomicCalculator.Basics.sortByEnergy — Method

Basics.sortByEnergy(multiplet::Hfs.HfMultiplet) ... to sort all hyperfine levels in the multiplet into a sequence of increasing energy; a (new) multiplet::Hfs.HfMultiplet is returned.

JenaAtomicCalculator.Basics.tabulate — Method

Basics.tabulate(stream::IO, ::HfEnergies, multiplet::Hfs.HfMultiplet) ... tabulates the energies of all hyperfine levels of the given multiplet; nothing is returned.

JenaAtomicCalculator.Basics.tabulate — Method

Basics.tabulate(stream::IO, ::HfEnergiesRelative, multiplet::Hfs.HfMultiplet) ... tabulates the hyperfine level energies relative to the lowest level of the multiplet; nothing is returned.

JenaAtomicCalculator.Hfs.amplitude — Method

Hfs.amplitude(mp::EmMultipole, rLevel::Level, sLevel::Level, grid::Radial.Grid; printout::Bool=true) ... to compute the T^(M1), T^(E2) or T^(M3) hyperfine amplitude <alphar Jr || T^(mp) || alphas Js> for a given pair of levels. A value::ComplexF64 is returned.

    Note (26-Jul-2026): coeff.T, for rank>0 one-particle operators, is returned by
    SpinAngular.computeCoefficientsNonScalar already in "GRASP-like" convention, i.e. with an internal
    factor sqrt(2*j_a+1) applied (see the active "GRASP like spin-angular coefficient" step in that
    function) -- this is NOT what the paper's Eq. (48)-type reduced-matrix-element sum
    <leftCsf||W^(k)||rightCsf> = sum_ab d^k_ab [n_a kappa_a||w^(k)||n_b kappa_b] needs (the "pure"
    coefficient, without that factor). By contrast, IsotopeShift.amplitude's rank-0 case uses
    SpinAngular.computeCoefficientsScalar, where the equivalent conversion step is deliberately commented
    out, and IsotopeShift.amplitude applies its own sqrt(2j_a+1) factor externally to match its formula.
    This function previously used coeff.T directly with no compensating division, leaving an
    uncorrected sqrt(2j_a+1) factor in every M1/E2/M3 hyperfine amplitude -- confirmed for H(1s) M1: the
    A-constant was too large by ~1.42 (= sqrt(2), the j=1/2 value) after the separate missing-alpha fix
    in InteractionStrength.hfs_tM1. Fixed by dividing each term by sqrt(2*j_a+1), undoing the internal
    GRASP-like conversion, matching IsotopeShift's convention.

JenaAtomicCalculator.Hfs.computeAmplitudesProperties — Method

`Hfs.computeAmplitudesProperties(outcome::Hfs.Outcome, nm::Nuclear.Model, grid::Radial.Grid, settings::Hfs.Settings, im::Hfs.InteractionMatrix) ... to compute all amplitudes and properties of for a given level; an outcome::Hfs.Outcome is returned for which the amplitudes and properties are evaluated explicitly.

JenaAtomicCalculator.Hfs.computeHyperfineMultiplet — Method

Hfs.computeHyperfineMultiplet(level::Level, nm::Nuclear.Model, grid::Radial.Grid) ... to compute a hyperfine multiplet, i.e. a representation of hyperfine levels within a hyperfine-coupled basis as defined by the given (electronic) level; a hfMultiplet::hfMultiplet is returned.

JenaAtomicCalculator.Hfs.computeHyperfineMultiplet — Method

Hfs.computeHyperfineMultiplet(multiplet::Multiplet, nm::Nuclear.Model, grid::Radial.Grid, settings::Hfs.Settings; output=true) ... to compute a hyperfine multiplet, i.e. a representation of hyperfine levels within a hyperfine-coupled basis as defined by the given (electronic) multiplet; a hfsMultiplet::IJF_Multiplet is returned.

JenaAtomicCalculator.Hfs.computeHyperfineRepresentation — Method

Hfs.computeHyperfineRepresentation(hfBasisVectors::Array{HfBasisVector,1}, nm::Nuclear.Model, grid::Radial.Grid) ... to set-up and diagonalized the Hamiltonian matrix of H^(DFB) + H^(hfs) within the atomic hyperfine (IJF-coupled) basis; a hfsMultiplet::IJF_Multiplet is returned.

JenaAtomicCalculator.Hfs.computeInteractionAmplitudeM — Method

Hfs.computeInteractionAmplitudeM(mp::EmMultipole, leftIsomer::Nuclear.Isomer, rightIsomer::Nuclear.Isomer) ... to compute the hyperfine interaction amplitude (<leftIsomer || M^(mp)) || rightIsomer>) for the interaction of two nuclear levels; this ME is geometrically fixed if the left and right isomer are the same, and it depends on the nuclear ME otherwise. An amplitude::ComplexF64 is returned.

JenaAtomicCalculator.Hfs.computeInteractionAmplitudeT — Method

Hfs.computeInteractionAmplitudeT(mp::EmMultipole, aLevel::Level, bLevel, grid::Radial.Grid) ... to compute the T^(mp) interaction amplitude for two levels of the same basis, i.e. (<aLevel || T^(mp) || bLevel>). A me::ComplexF64 is returned.

JenaAtomicCalculator.Hfs.computeInteractionMatrix — Method

Hfs.computeInteractionMatrix(basis::Basis, grid::Radial.Grid, settings::Hfs.Settings) ... to compute the T^M1 and/or T^E2 interaction matrices for the given basis, i.e. (<csfr || T^(n)) || csfs>). An im::Hfs.InteractionMatrix is returned.

    Note (26-Jul-2026): each coeff.T here is divided by sqrt(2*j_a+1), undoing the "GRASP-like"
    sqrt(2*j_a+1) factor that SpinAngular.computeCoefficientsNonScalar applies internally for rank>0
    one-particle operators -- see the matching note on Hfs.amplitude for the full explanation. Applies
    uniformly to the M1, E2, and M3 blocks below (all use the same rank>0 SpinAngular path).

JenaAtomicCalculator.Hfs.computeModifiedEinsteinRates — Method

Hfs.computeModifiedEinsteinRates(upperOutcome::Outcome, lowerOutcome::Outcome, multipoles::Array{EmMultipole,1}, gauge::EmGauge, grid::Radial.Grid) ... to compute and tabulate the modified Einstein amplitudes and rates for the hyperfine-resolved transitions between the upper and lower outcome. The procedures assumes that the two outcomes provide a proper IJF expansion (multiplet) of the hyperfine levels of interest. A neat table is printed but nothing is returned otherwise

JenaAtomicCalculator.Hfs.computeOutcomes — Method

Hfs.computeOutcomes(multiplet::Multiplet, nm::Nuclear.Model, grid::Radial.Grid, settings::Hfs.Settings; output=true) ... to compute (as selected) the HFS A, B and C parameters as well as hyperfine energy splittings for the levels of the given multiplet and as specified by the given settings. The results are printed in neat tables to screen and, if requested, an arrays{Hfs.Outcome,1} with all the results are returned.

JenaAtomicCalculator.Hfs.defineHyperfineBasis — Method

Hfs.defineHyperfineBasis(level::Level, nm::Nuclear.Model) ... to define/set-up an atomic hyperfine (IJF-coupled) basis for the given electronic level; a hfsBasis::IJF_Basis is returned.

JenaAtomicCalculator.Hfs.defineHyperfineBasis — Method

Hfs.defineHyperfineBasis(multiplet::Multiplet, nm::Nuclear.Model) ... to define/set-up an atomic hyperfine (IJF-coupled) basis for the given electronic multipet; a hfBasis::HfBasis is returned.

JenaAtomicCalculator.Hfs.determineOutcomes — Method

Hfs.determineOutcomes(multiplet::Multiplet, settings::Hfs.Settings) ... to determine a list of Outcomes's for the computation of HFS A- and B-parameters for the given multiplet. It takes into account the particular selections and settings. An Array{Hfs.Outcome,1} is returned. Apart from the level specification, all physical properties are set to zero during the initialization process.

JenaAtomicCalculator.Hfs.displayNondiagonal — Method

Hfs.displayNondiagonal(stream::IO, multiplet::Multiplet, grid::Radial.Grid, settings::Hfs.Settings) ... to compute and display all non-diagonal hyperfine amplitudes for the selected levels. A small neat table of all (pairwise) hyperfine amplitudes is printed but nothing is returned otherwise.

JenaAtomicCalculator.Hfs.displayOutcomes — Method

Hfs.displayOutcomes(outcomes::Array{Hfs.Outcome,1}) ... to display a list of levels that have been selected for the computations A small neat table of all selected levels and their energies is printed but nothing is returned otherwise.

JenaAtomicCalculator.Hfs.displayResults — Method

Hfs.displayResults(stream::IO, outcomes::Array{Hfs.Outcome,1}, nm::Nuclear.Model, settings::Hfs.Settings) ... to display the energies, A- and B-values, Delta E_F energy shifts, etc. for the selected levels. All nuclear parameters are taken from the nuclear model. A neat table is printed but nothing is returned otherwise.

Isotope shifts

JenaAtomicCalculator.IsotopeShift.AbstractIsotopeShiftKind — Type

abstract type IsotopeShift.AbstractIsotopeShiftKind ... selects the isotope-shift amplitude to be computed; used for typed dispatch in IsotopeShift.amplitude().

+ NMSamplitude        ... normal mass-shift amplitude        H^(NMS).
+ SMSamplitudeA       ... specific mass-shift amplitude A    H^(SMS,A).
+ SMSamplitudeB       ... specific mass-shift amplitude B    H^(SMS,B).
+ SMSamplitudeC       ... specific mass-shift amplitude C    H^(SMS,C).
+ FieldShiftAmplitude ... field-shift amplitude              H^(field-shift).
+ BosonFieldAmplitude ... boson-field shift amplitude        H^(boson-field).

JenaAtomicCalculator.IsotopeShift.Outcome — Type

struct IsotopeShift.Outcome ... defines a type to keep the outcome of a isotope-shift computation, such as the K and F parameters as well other results.

+ level                     ::Level              ... Atomic level to which the outcome refers to.
+ Knms                      ::Float64            ... K_nms parameter
+ Ksms                      ::Float64            ... K_sms parameter
+ Fme                       ::Float64            ... F parameter from matrix element.
+ Fdensity                  ::Float64            ... F parameter from density.
+ Xboson                    ::Float64            ... X boson-field shift constant.
+ amplitudeKnms             ::Complex{Float64}   ... K_nms amplitude
+ amplitudeKsmsA            ::Complex{Float64}   ... K_sms,A amplitude
+ amplitudeKsmsB            ::Complex{Float64}   ... K_sms,B amplitude
+ amplitudeKsmsC            ::Complex{Float64}   ... K_sms,C amplitude

JenaAtomicCalculator.IsotopeShift.Outcome — Method

IsotopeShift.Outcome() ... constructor for an empty instance of Hfs.Outcome for the computation of isotope-shift properties.

JenaAtomicCalculator.IsotopeShift.Settings — Type

struct IsotopeShift.Settings <: AbstractPropertySettings ... defines a type for the details and parameters of computing isotope-shift M and F parameters.

+ calcNMS                  ::Bool             ... True if mass-shift parameters M_nmn need to be calculated, and false otherwise.
+ calcSMS                  ::Bool             ... True if mass-shift parameters M_sms need to be calculated, and false otherwise.
+ calcF                    ::Bool             ... True if the field-shift parameter need to be calculated, and false otherwise.
+ calcBoson                ::Bool             ... True if the boson-field shift parameter need to be calculated, and false otherwise.
+ printBefore              ::Bool             ... True if a list of selected levels is printed before the actual computations start. 
+ bosonMass                ::Float64          ... mass of the scalar boson [e_electron].
+ levelSelection           ::LevelSelection   ... Specifies the selected levels, if any.

JenaAtomicCalculator.IsotopeShift.Settings — Method

IsotopeShift.Settings(set::IsotopeShift.Settings; keywords...) ... keyword constructor to modify selected fields of an IsotopeShift.Settings object; for fields not given, the values of set are used.

JenaAtomicCalculator.IsotopeShift.Settings — Method

IsotopeShift.Settings(; calcNMS::Bool=true, calcSMS::Bool=false, calcF::Bool=false, calcBoson::Bool=false, printBefore::Bool=true, bosonMass::Float64=0., levelSelection::LevelSelection=LevelSelection()) ... keyword constructor to overwrite selected value of isoshift computations.

JenaAtomicCalculator.IsotopeShift.amplitude — Method

IsotopeShift.amplitude(::BosonFieldAmplitude, rLevel::Level, sLevel::Level, potential::Array{Float64,1}, grid::Radial.Grid) ... to compute the boson-field shift amplitude <alphar Jr || H^(boson-field) || alphas Js> for a given pair of levels. The potential must provide the effective boson-field potential. A value::ComplexF64 is returned.

JenaAtomicCalculator.IsotopeShift.amplitude — Method

IsotopeShift.amplitude(::FieldShiftAmplitude, rLevel::Level, sLevel::Level, potential::Array{Float64,1}, grid::Radial.Grid) ... to compute the field-shift amplitude <alphar Jr || H^(field-shift) || alphas Js> for a given pair of levels. The potential must provide delta V^(nuc). A value::ComplexF64 is returned.

JenaAtomicCalculator.IsotopeShift.amplitude — Method

IsotopeShift.amplitude(::NMSamplitude, rLevel::Level, sLevel::Level, nm::Nuclear.Model, grid::Radial.Grid) ... to compute the normal mass-shift amplitude <alphar Jr || H^(NMS) || alphas Js> for a given pair of levels. A value::ComplexF64 is returned.

JenaAtomicCalculator.IsotopeShift.amplitude — Method

IsotopeShift.amplitude(::SMSamplitudeA, rLevel::Level, sLevel::Level, nm::Nuclear.Model, grid::Radial.Grid) ... to compute the specific mass-shift amplitude A <alphar Jr || H^(SMS,A) || alphas Js> for a given pair of levels. A value::ComplexF64 is returned.

JenaAtomicCalculator.IsotopeShift.amplitude — Method

IsotopeShift.amplitude(::SMSamplitudeB, rLevel::Level, sLevel::Level, nm::Nuclear.Model, grid::Radial.Grid) ... to compute the specific mass-shift amplitude B <alphar Jr || H^(SMS,B) || alphas Js> for a given pair of levels. A value::ComplexF64 is returned.

JenaAtomicCalculator.IsotopeShift.amplitude — Method

IsotopeShift.amplitude(::SMSamplitudeC, rLevel::Level, sLevel::Level, nm::Nuclear.Model, grid::Radial.Grid) ... to compute the specific mass-shift amplitude C <alphar Jr || H^(SMS,C) || alphas Js> for a given pair of levels. A value::ComplexF64 is returned.

JenaAtomicCalculator.IsotopeShift.computeAmplitudesProperties — Method

`IsotopeShift.computeAmplitudesProperties(outcome::IsotopeShift.Outcome, nm::Nuclear.Model, grid::Radial.Grid, settings::IsotopeShift.Settings) ... to compute all amplitudes and properties for a given level; an outcome::IsotopeShift.Outcome is returned for which the amplitudes and properties are now evaluated explicitly.

JenaAtomicCalculator.IsotopeShift.computeOutcomes — Method

IsotopeShift.computeOutcomes(multiplet::Multiplet, nm::Nuclear.Model, grid::Radial.Grid, settings::IsotopeShift.Settings; output=true) ... to compute (as selected) the isotope-shift M and F parameters for the levels of the given multiplet and as specified by the given settings. The results are printed in neat tables to screen but nothing is returned otherwise.

JenaAtomicCalculator.IsotopeShift.determineOutcomes — Method

IsotopeShift.determineOutcomes(multiplet::Multiplet, settings::IsotopeShift.Settings) ... to determine a list of Outcomes's for the computation of the isotope-shift M and F parameters for the given multiplet. It takes into account the particular selections and settings. An Array{IsotopeShift.Outcome,1} is returned. Apart from the level specification, all physical properties are set to zero during the initialization process.

JenaAtomicCalculator.IsotopeShift.displayOutcomes — Method

IsotopeShift.displayOutcomes(outcomes::Array{IsotopeShift.Outcome,1}) ... to display a list of levels that have been selected for the computations. A small neat table of all selected levels and their energies is printed but nothing is returned otherwise.

JenaAtomicCalculator.IsotopeShift.displayResults — Method

IsotopeShift.displayResults(stream::IO, outcomes::Array{Hfs.Outcome,1}, settings::IsotopeShift.Settings) ... to display the energies, M_ms and F-parameters, etc. for the selected levels. A neat table is printed but nothing is returned otherwise.